requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.301
langsmith==0.0.40
numpy==1.26.0
openai==0.28.1
pandas==1.5.2
python-dotenv==0.21.0
python-multipart==0.0.6
requests==2.28.1
pypdf==3.16.2
chromadb==0.4.13
